fix: refactoring agenticChatController.ts
