+++ server/aws-lsp-codewhisperer/src/language-server/agenticChat/agenticChatController.ts
@@ -129,6 +129,7 @@ import { URI } from 'vscode-uri'
 import { McpManager } from './tools/mcp/mcpManager'
 import { McpTool } from './tools/mcp/mcpTool'
 import { processMcpToolUseMessage } from './tools/mcp/mcpUtils'
+import { McpUiController } from './tools/mcp/mcpUiController'
 import { CommandCategory } from './tools/executeBash'
 import { UserWrittenCodeTracker } from '../../shared/userWrittenCodeTracker'
 
@@ -160,6 +161,7 @@ export class AgenticChatController implements ChatHandlers {
     #contextCommandsProvider: ContextCommandsProvider
     #stoppedToolUses = new Set<string>()
     #userWrittenCodeTracker: UserWrittenCodeTracker | undefined
+    #mcpUiController: McpUiController
 
     /**
      * Determines the appropriate message ID for a tool use based on tool type and name
@@ -194,6 +196,7 @@ export class AgenticChatController implements ChatHandlers {
             this.#features.workspace,
             this.#features.lsp
         )
+        this.#mcpUiController = new McpUiController(features)
     }
 
     async onButtonClick(params: ButtonClickParams): Promise<ButtonClickResult> {
@@ -365,137 +368,11 @@ export class AgenticChatController implements ChatHandlers {
     }
 
     async onListMcpServers(params: ListMcpServersParams) {
-        const mcpManagerServerConfigs = McpManager.instance.getAllServerConfigs()
-        const serversAndTools = McpManager.instance.listServersAndTools()
-
-        // Transform server configs into DetailedListItem objects
-        const activeItems: DetailedListItem[] = []
-        const disabledItems: DetailedListItem[] = []
-
-        Array.from(mcpManagerServerConfigs.entries()).forEach(([serverName, config]) => {
-            const toolNames = serversAndTools[serverName] || []
-            const toolsCount = toolNames.length
-
-            const item: DetailedListItem = {
-                title: serverName,
-                description: `Command: ${config.command}`,
-            }
-
-            if (config.disabled) {
-                disabledItems.push(item)
-            } else {
-                activeItems.push(item)
-            }
-        })
-
-        // Create the groups
-        const groups: DetailedListGroup[] = []
-
-        if (activeItems.length > 0) {
-            groups.push({
-                groupName: 'Active',
-                children: activeItems,
-            })
-        }
-
-        if (disabledItems.length > 0) {
-            groups.push({
-                groupName: 'Disabled',
-                children: disabledItems,
-            })
-        }
-
-        // Return the result in the expected format
-        return {
-            header: {
-                title: 'MCP Servers',
-                description:
-                    "Q automatically uses any MCP servers that have been added, so you don't have to add them as context.",
-            },
-            list: groups,
-        }
+        return this.#mcpUiController.onListMcpServers()
     }
 
     async onMcpServerClick(params: McpServerClickParams): Promise<McpServerClickResult> {
-        this.#log(`[VSCode Server] onMcpServerClick event with params: ${JSON.stringify(params)}`)
-        if (params.id === 'add-new-mcp') {
-            return {
-                id: params.id,
-                header: {
-                    title: 'Add MCP Server',
-                    status: {},
-                    actions: [],
-                },
-                list: [],
-                filterActions: [
-                    {
-                        id: 'cancel-mcp',
-                        text: 'Cancel',
-                    },
-                    {
-                        id: 'save-mcp',
-                        text: 'Save',
-                        status: 'primary',
-                    },
-                ],
-                filterOptions: [
-                    {
-                        type: 'radiogroup',
-                        id: 'scope',
-                        title: 'Scope',
-                        options: [
-                            {
-                                label: `Global - Used globally. Edit config`,
-                                value: 'Yes',
-                            },
-                            {
-                                label: `This workspace - Only used in this workspace. Edit config`,
-                                value: 'No',
-                            },
-                        ],
-                    },
-                    {
-                        type: 'textinput', // User input text
-                        id: 'name',
-                        title: 'Name',
-                    },
-                    {
-                        type: 'select', // Drop down
-                        id: 'transport',
-                        title: 'Transport',
-                        options: [
-                            {
-                                label: 'stdio',
-                                value: 'yes',
-                            },
-                        ],
-                    },
-                    {
-                        type: 'textinput', // User input text
-                        id: 'command',
-                        title: 'Command',
-                    },
-                    {
-                        type: 'numericinput',
-                        id: 'timeout',
-                        title: 'Timeout',
-                        description: 'Seconds',
-                    },
-                ],
-            }
-        } else if (params.id === 'open-mcp-xx') {
-            // Edit
-        }
-        return {
-            id: params.id,
-            header: {
-                title: 'MCP Servers',
-                status: {},
-                description:
-                    'Q automatically uses any MCP servers that have been added, so you don\'t have to add them as context. All MCPs are defaulted to "Ask before running".',
-                actions: [],
-            },
-        }
+        return this.#mcpUiController.onMcpServerClick(params)
     }
 
     async #sendProgressToClient(chunk: ChatResult | string, partialResultToken?: string | number) {
